rename namespace Immerse.BfhClient -> Immerse.BfhClient.Game
