+++ Assets/Scripts/Game/ActionPopup.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using UnityEngine;
 
-namespace Immerse.BfhClient
+namespace Immerse.BfhClient.Game
 {
     public class ActionPopup : MonoBehaviour
     {
