+++ Assets/Scripts/Game/Country.cs
@@ -2,12 +2,12 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Immerse.BfhClient
+namespace Immerse.BfhClient.Game
 {
     [System.Serializable]
     public class Country
     {
-        [SerializeField] public string Name { get; private set; }
+        public string Name { get; private set; }
 
         public Country(string name)
         {
