+++ Assets/Scripts/Game/GameCanvasController.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Immerse.BfhClient
+namespace Immerse.BfhClient.Game
 {
     public class GameCanvasController : MonoBehaviour
     {
