+++ Assets/Scripts/Game/RegionSelectorDebugger.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace Immerse.BfhClient
+namespace Immerse.BfhClient.Game
 {
 	public class RegionSelectorDebugger : MonoBehaviour
 	{
