+++ Assets/Scripts/Game/SerializableRegion.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Immerse.BfhClient
+namespace Immerse.BfhClient.Game
 {
 	[Serializable]
 	public class SerializableRegion
